# docs/swagger/swagger-sdce-1.json
{
  "openapi" : "3.0.1",
  "info" : {
    "contact" : {
      "email" : "onap-discuss@lists.onap.org",
      "name" : "ONAP",
      "url" : "https://onap.readthedocs.io"
    },
    "description" : "SDC API for onboarding (SDCE-1)",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0"
    },
    "title" : "SPC API: SDCE-1",
    "version" : "1.0"
  },
  "servers" : [ {
    "description" : "SDCE-1 APIs",
    "url" : "/sdc"
  } ],
  "paths" : {
...
    "/v1.0/vendor-software-products/{vspId}/versions/{versionId}/components/{componentId}/compute-flavors/{computeFlavorId}/questionnaire" : {
      "get" : {
        "description" : "Get vendor software product component compute-flavor questionnaire",
        "operationId" : "getQuestionnaire_1",
        "parameters" : [ {
          "description" : "Vendor software product Id",
          "in" : "path",
          "name" : "vspId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Version Id",
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor software product component Id",
          "in" : "path",
          "name" : "componentId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor software product compute-flavor Id",
          "in" : "path",
          "name" : "computeFlavorId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/QuestionnaireResponseDto"
                }
              }
            }
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor Software Product Component Compute-flavors" ]
      },
      "put" : {
        "description" : "Update vendor software product component compute-flavor questionnaire",
        "operationId" : "updateQuestionnaire_1",
        "parameters" : [ {
          "description" : "Vendor software product Id",
          "in" : "path",
          "name" : "vspId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Version Id",
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor software product component Id",
          "in" : "path",
          "name" : "componentId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor software product compute-flavor Id",
          "in" : "path",
          "name" : "computeFlavorId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "string"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : { }
            },
            "description" : "default response"
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor Software Product Component Compute-flavors" ]
      }
    },
...
    "/v1.0/vendor-software-products/{vspId}/versions/{versionId}/components/{componentId}/images/{imageId}/questionnaire" : {
      "get" : {
        "description" : "Get vendor software product component image questionnaire",
        "operationId" : "getQuestionnaire_2",
        "parameters" : [ {
          "description" : "Vendor software product Id",
          "in" : "path",
          "name" : "vspId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Version Id",
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor software product component Id",
          "in" : "path",
          "name" : "componentId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor software product image Id",
          "in" : "path",
          "name" : "imageId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/QuestionnaireResponseDto"
                }
              }
            }
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor Software Product Images" ]
      },
      "put" : {
        "description" : "Update vendor software product component image questionnaire",
        "operationId" : "updateQuestionnaire_2",
        "parameters" : [ {
          "description" : "Vendor software product Id",
          "in" : "path",
          "name" : "vspId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Version Id",
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor software product component Id",
          "in" : "path",
          "name" : "componentId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor software product image Id",
          "in" : "path",
          "name" : "imageId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "string"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : { }
            },
            "description" : "default response"
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor Software Product Images" ]
      }
    },
...
    "/v1.0/vendor-software-products/{vspId}/versions/{versionId}/components/{componentId}/nics/{nicId}/questionnaire" : {
      "get" : {
        "description" : "Get vendor software product component NIC questionnaire",
        "operationId" : "getQuestionnaire_3",
        "parameters" : [ {
          "description" : "Vendor software product Id",
          "in" : "path",
          "name" : "vspId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor software product version Id",
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor software product component Id",
          "in" : "path",
          "name" : "componentId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor software product NIC Id",
          "in" : "path",
          "name" : "nicId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/QuestionnaireResponseDto"
                }
              }
            }
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor Software Product Component NICs" ]
      },
      "put" : {
        "description" : "Update vendor software product component NIC questionnaire",
        "operationId" : "updateQuestionnaire_3",
        "parameters" : [ {
          "description" : "Vendor software product Id",
          "in" : "path",
          "name" : "vspId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor software product version Id",
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor software product component Id",
          "in" : "path",
          "name" : "componentId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor software product NIC Id",
          "in" : "path",
          "name" : "nicId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "string"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : { }
            },
            "description" : "default response"
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor Software Product Component NICs" ]
      }
    },
...
    "/v1.0/vendor-software-products/{vspId}/versions/{versionId}/components/{componentId}/questionnaire" : {
      "get" : {
        "description" : "Get vendor software product component questionnaire",
        "operationId" : "getQuestionnaire",
        "parameters" : [ {
          "description" : "Vendor software product Id",
          "in" : "path",
          "name" : "vspId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Version Id",
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor software product component Id",
          "in" : "path",
          "name" : "componentId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/QuestionnaireResponseDto"
                }
              }
            }
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor Software Product Components" ]
      },
      "put" : {
        "description" : "Update vendor software product component questionnaire",
        "operationId" : "updateQuestionnaire",
        "parameters" : [ {
          "description" : "Vendor software product Id",
          "in" : "path",
          "name" : "vspId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Version Id",
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor software product component Id",
          "in" : "path",
          "name" : "componentId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "string"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : { }
            },
            "description" : "default response"
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor Software Product Components" ]
      }
    },
...
    "/v1.0/vendor-software-products/{vspId}/versions/{versionId}/questionnaire" : {
      "get" : {
        "description" : "Get vendor software product questionnaire",
        "operationId" : "getQuestionnaire_4",
        "parameters" : [ {
          "in" : "path",
          "name" : "vspId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/QuestionnaireResponseDto"
                }
              }
            }
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor Software Products" ]
      },
      "put" : {
        "description" : "Update vendor software product questionnaire",
        "operationId" : "updateQuestionnaire_4",
        "parameters" : [ {
          "in" : "path",
          "name" : "vspId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "string"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : { }
            },
            "description" : "default response"
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor Software Products" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ErrorMessage" : {
        "type" : "object",
        "properties" : {
          "level" : {
            "type" : "string",
            "enum" : [ "ERROR", "WARNING", "INFO" ]
          },
          "message" : {
            "type" : "string"
          }
        }
      },
...
      "QuestionnaireResponseDto" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "type" : "string"
          },
          "errorMessage" : {
            "$ref" : "#/components/schemas/ErrorMessage"
          },
          "schema" : {
            "type" : "string"
          }
        }
      },
...
    }
  }
}